backend/requirements.txt
Flask
gunicorn
pandas
SQLAlchemy
flask_cors
flask_sqlalchemy
flask_migrate
config
mysql-connector-python
PyMySQL
python-dotenv
boto3